+++ lib/esp32/device.ex
@@ -9,8 +9,21 @@ defmodule Esp32.Device do
           chip: Esp32.Chip.name() | nil,
           baud: pos_integer(),
           stub?: boolean(),
-          usb_otg?: boolean()
+          usb_otg?: boolean(),
+          reset: Esp32.Reset.strategy(),
+          reset_pin: term(),
+          boot_pin: term()
         }
 
-  defstruct [:uart, :port, :chip, baud: 115_200, stub?: false, usb_otg?: false]
+  defstruct [
+    :uart,
+    :port,
+    :chip,
+    :reset_pin,
+    :boot_pin,
+    baud: 115_200,
+    stub?: false,
+    usb_otg?: false,
+    reset: :none
+  ]
 end